requirements.txt
python-telegram-bot
google-generativeai
python-dotenv
requests
supabase
